Revert "Update battle_main.h"
This reverts commit a84a77dc3de95bc7f54d5f03d48862739fb03587.

+++ include/battle_main.h
@@ -10,7 +10,7 @@ struct TrainerMoney
 struct TrainerBall
 {
     u8 classId;
-    u16 Ball;
+    u8 Ball;
 };
 
 struct UnknownPokemonStruct4
